+++ packages/flutter_bloc_one_shot/lib/flutter_bloc_one_shot.dart
@@ -5,5 +5,6 @@
 library flutter_bloc_one_shot;
 
 export 'package:bloc_one_shot/bloc_one_shot.dart';
+export 'src/multiple_side_effect_listener.dart';
 export 'src/side_effect_consumer.dart';
 export 'src/side_effect_listener.dart';