+++ packages/flutter_bloc_one_shot/CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog
+
+## 0.2.0
+
+- Add `MultipleSideEffectListener` widget to merge multiple `SideEffectListener` widgets without nesting.
+- Refactor `SideEffectListener` to extend `SingleChildStatefulWidget` for compatibility with `MultipleSideEffectListener`.
+- Add `provider` as a direct dependency.
+
 ## 0.1.0
 
 - Initial release.